% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{getpareto}
\alias{getpareto}
\title{Find the pareto set}
\usage{
getpareto(scores)
}
\arguments{
\item{scores}{a matrix of negatively oriented scores where the rows correspond to different forecasts and the columns denote different scores.}
}
\value{
a vector of indices indicating all members of the pareto set.
}
\description{
Determine the set of pareto optimal forecasts in a matrix of scores
}
\details{
The Pareto set contains all those forecasts for which no other forecast is better in every respect. In this function, we assume that all scores are negatively oriented, "better" therefore means lower values.
}
\note{
This function becomes very memory hungry if you have more than 1000 forecasts, be careful.
}
