% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{hemd}
\alias{hemd}
\title{histogram emd}
\usage{
hemd(h1, h2, mids = NULL)
}
\arguments{
\item{h1, h2}{vectors of non-negtaive numbers representing two histograms}

\item{mids}{the bin mids corresponding to the histograms. Can also be given via the names of \code{h1}.}
}
\value{
the value of the EMD
}
\description{
Earth Mover's Distance between two histograms, given as vectors
}
