% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowlevel.R
\docType{data}
\name{raincols}
\alias{raincols}
\title{rain color scale}
\format{An object of class \code{character} of length 8.}
\usage{
raincols
}
\description{
eight shades of blue used in \code{plot.sadforecast}
}
\keyword{datasets}
