% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rrain}
\alias{rrain}
\title{Random Rain}
\format{A \code{4x10x128x128} matrix}
\source{
simulated using the 'RandomFields' package, code available at <10.5281/zenodo.3257511>
}
\usage{
data(rrain)
}
\description{
Randomly simulated synthetic rain fields with Matern covariances
}
\details{
These fields were used in Buschow et al. (2019) <doi:10.5194/gmd-12-3401-2019>. The first array corresponds to the four model configurations from that paper (different roughness nu and scale sc), the second dimension contains ten realizations for each model.
}
\examples{
data(rrain)
}
\keyword{datasets}
