% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{semd}
\alias{semd}
\title{spectral emd}
\usage{
semd(dt1, dt2, a = 1, dir = TRUE)
}
\arguments{
\item{dt1, dt2}{forecast and observed spectrum}

\item{a}{ratio between scale- and directional component}

\item{dir}{whether or not to include direction information}
}
\value{
a single value, the emd. If \code{dir=FALSE}, the value is signed, indicating the direction of the scale error.
}
\description{
earth mover's distance between dual-tree wavelet spectra
}
